+++ dst/app/middlewares/notFoundHandler.js
@@ -4,7 +4,7 @@ exports.notFoundHandler = void 0;
 /**
  * 404ハンドラーミドルウェア
  */
-const domain_1 = require("@cinerino/domain");
+const domain_1 = require("@chevre/domain");
 function notFoundHandler(req, __, next) {
     next(new domain_1.chevre.factory.errors.NotFound(`router for [${req.originalUrl}]`));
 }